V12.4更新版本

+++ WaterCloud.Service/SystemSecurity/LogService.cs
@@ -1,4 +1,4 @@
-﻿/*******************************************************************************
+/*******************************************************************************
  * Copyright © 2020 WaterCloud.Framework 版权所有
  * Author: WaterCloud
  * Description: WaterCloud快速开发平台
@@ -20,6 +20,7 @@ namespace WaterCloud.Service.SystemSecurity
     {
         //登录信息保存方式
         private string LoginProvider = GlobalContext.SystemConfig.LoginProvider;
+        private string HandleLogProvider = GlobalContext.SystemConfig.HandleLogProvider;
         private ModuleService moduleservice;
         //获取类名
         private string className = System.Reflection.MethodBase.GetCurrentMethod().DeclaringType.FullName.Split('.')[3];
@@ -27,14 +28,10 @@ namespace WaterCloud.Service.SystemSecurity
         {
             moduleservice = new ModuleService(context);
         }
-        public async Task<List<LogEntity>> GetLookList(Pagination pagination, int timetype, string keyword="")
+        public async Task<List<LogEntity>> GetList(Pagination pagination, int timetype, string keyword="")
         {
             //获取数据权限
-            var list = GetDataPrivilege("u", className.Substring(0, className.Length - 7));
-            if (!string.IsNullOrEmpty(keyword))
-            {
-                list = list.Where(u => u.F_Account.Contains(keyword) || u.F_Description.Contains(keyword) || u.F_ModuleName.Contains(keyword));
-            }
+            var result = new List<LogEntity>();
             DateTime startTime = DateTime.Now.ToString("yyyy-MM-dd").ToDate();
             DateTime endTime = DateTime.Now.ToString("yyyy-MM-dd").ToDate().AddDays(1);
             switch (timetype)
@@ -53,12 +50,44 @@ namespace WaterCloud.Service.SystemSecurity
                 default:
                     break;
             }
-            list = list.Where(t => t.F_Date >= startTime && t.F_Date <= endTime);
-            return GetFieldsFilterData(await repository.OrderList(list, pagination), className.Substring(0, className.Length - 7));
+            if (HandleLogProvider != Define.CACHEPROVIDER_REDIS)
+            {
+                var list = repository.IQueryable();
+                if (!string.IsNullOrEmpty(keyword))
+                {
+                    list = list.Where(u => u.F_Account.Contains(keyword) || u.F_Description.Contains(keyword) || u.F_ModuleName.Contains(keyword));
+                }
+
+                list = list.Where(t => t.F_Date >= startTime && t.F_Date <= endTime);
+                result = await repository.OrderList(list, pagination);
+            }
+            else
+            {
+                result = HandleLogHelper.HGetAll<LogEntity>(currentuser.CompanyId).Values.ToList();
+                if (!string.IsNullOrEmpty(keyword))
+                {
+                    result = result.Where(u => u.F_Account.Contains(keyword) || u.F_Description.Contains(keyword) || u.F_ModuleName.Contains(keyword)).Where(t => t.F_Date >= startTime && t.F_Date <= endTime).ToList();
+                }
+                else
+                {
+                    result = result.Where(t => t.F_Date >= startTime && t.F_Date <= endTime).ToList();
+                }
+                pagination.records = result.Count();
+                result = result.OrderByDescending(a => a.F_CreatorTime).Skip((pagination.page - 1) * pagination.rows).Take(pagination.rows).ToList();
+
+            }
+            return GetFieldsFilterData(result, className.Substring(0, className.Length - 7));
         }
         public async Task<List<LogEntity>> GetList()
-        {           
-            return repository.IQueryable().ToList();
+        {
+            if (HandleLogProvider != Define.CACHEPROVIDER_REDIS)
+            {
+                return repository.IQueryable().ToList();
+            }
+            else
+            {
+                return HandleLogHelper.HGetAll<LogEntity>(currentuser.CompanyId).Values.ToList(); ;
+            }
         }
         public async Task RemoveLog(string keepTime)
         {
@@ -75,9 +104,18 @@ namespace WaterCloud.Service.SystemSecurity
             {
                 operateTime = DateTime.Now.AddMonths(-3);
             }
-            var expression = ExtLinq.True<LogEntity>();
-            expression = expression.And(t => t.F_Date <= operateTime);
-            await repository.Delete(expression);
+            if (HandleLogProvider != Define.CACHEPROVIDER_REDIS)
+            {
+                var expression = ExtLinq.True<LogEntity>();
+                expression = expression.And(t => t.F_Date <= operateTime);
+                await repository.Delete(expression);
+            }
+            else
+            {
+                var list = HandleLogHelper.HGetAll<LogEntity>(currentuser.CompanyId).Values.ToList();
+                var strList = list.Where(t => t.F_Date <= operateTime).Select(a=>a.F_Id).ToList();
+                await HandleLogHelper.HDelAsync(currentuser.CompanyId, strList.ToArray());
+            }
         }
         public async Task WriteDbLog(bool result, string resultLog)
         {
@@ -92,7 +130,14 @@ namespace WaterCloud.Service.SystemSecurity
             logEntity.F_Result = result;
             logEntity.F_Description = resultLog;
             logEntity.Create();
-            await repository.Insert(logEntity);
+            if (HandleLogProvider != Define.CACHEPROVIDER_REDIS)
+            {
+                await repository.Insert(logEntity);
+            }
+            else
+            {
+                await HandleLogHelper.HSetAsync(currentuser.CompanyId, logEntity.F_Id, logEntity);
+            }
         }
         public async Task WriteDbLog(LogEntity logEntity)
         {
@@ -104,7 +149,7 @@ namespace WaterCloud.Service.SystemSecurity
                 {
                     logEntity.F_IPAddress = LoginProvider=="WebApi"? "未连接未知": WebHelper.Ip;
                     logEntity.F_IPAddressName = "本地局域网";
-                    logEntity.F_CompanyId = Define.SYSTEM_MASTERPROJECT;
+                    logEntity.F_CompanyId = GlobalContext.SystemConfig.SysemMasterProject;
                 }
                 else
                 {
@@ -113,23 +158,67 @@ namespace WaterCloud.Service.SystemSecurity
                     logEntity.F_CompanyId = currentuser.CompanyId;
                 }
                 logEntity.Create();
-                await repository.Insert(logEntity);
+                if (HandleLogProvider != Define.CACHEPROVIDER_REDIS)
+                {
+                    uniwork.Rollback();
+                    await repository.Insert(logEntity);
+                }
+                else
+                {
+                    await HandleLogHelper.HSetAsync(logEntity.F_CompanyId, logEntity.F_Id, logEntity);
+                }
             }
             catch (Exception)
             {
                 logEntity.F_IPAddress = LoginProvider == "WebApi" ? "未连接未知" : WebHelper.Ip;
                 logEntity.F_IPAddressName = "本地局域网";
-                logEntity.F_CompanyId = Define.SYSTEM_MASTERPROJECT;
+                logEntity.F_CompanyId = GlobalContext.SystemConfig.SysemMasterProject;
                 logEntity.Create();
-                await repository.Insert(logEntity);
+                if (HandleLogProvider != Define.CACHEPROVIDER_REDIS)
+                {
+                    await repository.Insert(logEntity);
+                }
+                else
+                {
+                    await HandleLogHelper.HSetAsync(logEntity.F_CompanyId, logEntity.F_Id, logEntity);
+                }
             }
         }
 
+        public async Task<LogEntity> CreateLog(string className, DbLogType type)
+        {
+            try
+            {
+                var moduleitem = (await moduleservice.GetList()).Where(a => a.F_IsExpand == false && a.F_EnCode == className.Substring(0, className.Length - 10)).FirstOrDefault();
+                if (moduleitem==null)
+                {
+                    throw new Exception();
+                }
+                var module = (await moduleservice.GetList()).Where(a => a.F_Id == moduleitem.F_ParentId).FirstOrDefault();
+                return new LogEntity(await CreateModule(module), moduleitem == null ? "" : moduleitem.F_FullName, type.ToString());
+            }
+            catch (Exception)
+            {
+                return new LogEntity(className, "" , type.ToString());
+            }
+        }
         public async Task<LogEntity> CreateLog(string className, string type)
         {
-            var moduleitem = (await moduleservice.GetList()).Where(a => a.F_IsExpand == false && a.F_EnCode == className.Substring(0, className.Length - 10)).FirstOrDefault();
-            var module = (await moduleservice.GetList()).Where(a =>  a.F_Id == moduleitem.F_ParentId).FirstOrDefault();
-            return new LogEntity(await CreateModule(module), moduleitem==null? "": moduleitem.F_FullName, type);
+            try
+            {
+                var moduleitem = (await moduleservice.GetList()).Where(a => a.F_IsExpand == false && a.F_EnCode == className.Substring(0, className.Length - 10)).FirstOrDefault();
+                if (moduleitem == null)
+                {
+                    throw new Exception();
+                }
+                var module = (await moduleservice.GetList()).Where(a => a.F_Id == moduleitem.F_ParentId).FirstOrDefault();
+                return new LogEntity(await CreateModule(module), moduleitem == null ? "" : moduleitem.F_FullName, type);
+
+            }
+            catch (Exception)
+            {
+                return new LogEntity(className, "", type.ToString());
+            }
         }
         public async Task<string> CreateModule(ModuleEntity module, string str="")
         {
@@ -147,6 +236,44 @@ namespace WaterCloud.Service.SystemSecurity
                 var temp= (await moduleservice.GetList()).Where(a =>a.F_Id==module.F_ParentId).FirstOrDefault();
                 return await CreateModule(temp ,str);
             }
+        }
+        public async Task WriteLog(string message, string className, string keyValue = "", DbLogType? logType = null, bool isError = false)
+        {
+            LogEntity logEntity;
+            if (logType != null)
+            {
+                logEntity = await CreateLog(className, (DbLogType)logType);
+                logEntity.F_Description += logType.ToDescription();
+            }
+            else
+            {
+                if (string.IsNullOrEmpty(keyValue))
+                {
+                    logEntity = await CreateLog(className, DbLogType.Create);
+                    logEntity.F_Description += DbLogType.Create.ToDescription();
+                }
+                else
+                {
+                    logEntity = await CreateLog(className, DbLogType.Update);
+                    logEntity.F_Description += DbLogType.Update.ToDescription();
+                }
+            }
+            logEntity.F_KeyValue = keyValue;
+            if (isError)
+            {
+                logEntity.F_Result = false;
+                logEntity.F_Description += "操作失败，" + message;
+            }
+            else
+            {
+                logEntity.F_Description += message;
+            }
+            if (currentuser != null && currentuser.UserId != null)
+            {
+                logEntity.F_Account = currentuser.UserCode;
+                logEntity.F_NickName = currentuser.UserName;
+            }
+            await WriteDbLog(logEntity);
         }
     }
 }
